% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_equal_grid.R
\name{st_equal_grid}
\alias{st_equal_grid}
\title{Create and equal area grid for multiple maps}
\usage{
st_equal_grid(places, titles, buffer, map_theme)
}
\arguments{
\item{places}{a list of sf object}

\item{titles}{a character vector of title for the maps}

\item{buffer}{a numeric vector of buffer values}

\item{map_theme}{a ggplot theme}
}
\value{
A list of sf objects
}
\description{
Create and equal area grid for multiple maps
}
\keyword{map}
\keyword{sf}
\keyword{spatial}
