% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_transform_outline.R
\name{st_transform_outline}
\alias{st_transform_outline}
\title{Create outline of sf projection}
\usage{
st_transform_outline(crs)
}
\arguments{
\item{crs}{The crs for transforming the sf}
}
\value{
An sf object
}
\description{
Create outline of sf projection
}
\keyword{map}
\keyword{sf}
\keyword{spatial}
