% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_transform_repair.R
\name{st_transform_repair}
\alias{st_transform_repair}
\title{Repair sf polygons}
\usage{
st_transform_repair(x, crs)
}
\arguments{
\item{x}{An sf object with unprojected coordinates}

\item{crs}{The crs for transforming the sf}
}
\value{
An sf object
}
\description{
Repair sf polygons
}
\keyword{map}
\keyword{meridian}
\keyword{sf}
\keyword{spatial}
